Refactor views directory structure
